rename: move repository folder
