+++ src/main/java/discodeit/repository/ChannelRepository.java
@@ -1,4 +1,4 @@
-package discodeit.service.repository;
+package discodeit.repository;
 
 import discodeit.entity.Channel;
