+++ src/main/java/discodeit/repository/MessageRepository.java
@@ -1,4 +1,4 @@
-package discodeit.service.repository;
+package discodeit.repository;
 
 
 import discodeit.entity.Message;
@@ -8,7 +8,10 @@ import java.util.Map;
 
 public interface MessageRepository {
     void save(Message message);
+
     Message loadById(String messageId);
+
     Map<String, Message> loadAll();
+
     void delete(Message message);
 }
